requirements.txt
jinja2
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
google-genai
python-docx
SpeechRecognition
PyAudio
